don't try setting the framerate if we're using a mode that does not require it
The _fps property can be used both to setup the camera and to setup
a timestamp estimator. When in a fixed trigger mode, both values have
to match (and therefore the _fps has to match a framerate that the
camera supports). When in a e.g. a hardware trigger mode, however,
the _fps value does *not* have to be supported by the camera.

So, allow arbitrary _fps parameter when in non-fixed trigger mode
so that timestamp estimators are configured properly.